requirements.txt
pyspark==3.5.1
pandas==2.2.2
numpy==1.26.4
seaborn==0.13.2
matplotlib==3.8.4
streamlit==1.36.0
fsspec==2024.3.1
python-dotenv==1.0.1
